requirements.txt
acme==1.1.0
appdirs==1.4.4
apturl==0.5.2
argcomplete==1.12.1
astroid==2.4.2
attrs==19.3.0
autopep8==1.5.3
backcall==0.1.0
bcrypt==3.1.7
beautifulsoup4==4.8.2
bleach==3.1.5
blinker==1.4
Brlapi==0.7.0
cachetools==4.2.2
certbot==0.40.0
certifi==2019.11.28
chardet==3.0.4
chrome-gnome-shell==0.0.0
Click==7.0
click-plugins==1.1.1
colorama==0.4.3
colour==0.1.5
command-not-found==0.3
ConfigArgParse==0.13.0
configobj==5.0.6
coverage==5.1
cryptography==2.8
cupshelpers==1.0
cycler==0.10.0
Cython==0.29.21
dateparser==1.0.0
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
defusedxml==0.6.0
dill==0.3.3
distlib==0.3.1
distro==1.4.0
distro-info===0.23ubuntu1
dlx==1.0.4
docplex==2.19.202
duplicity==0.8.12.0
entrypoints==0.3
fastdtw==0.3.4
fasteners==0.14.1
fastjsonschema==2.14.5
filelock==3.0.12
flake8==3.8.3
Flask==1.1.1
future==0.18.2
gitdb==4.0.5
GitPython==3.1.3
google-api-core==1.26.3
google-api-python-client==2.4.0
google-auth==1.30.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
googleapis-common-protos==1.53.0
gpg===1.13.1-unknown
gspread==3.7.0
h5py==3.1.0
html5lib==1.0.1
htmldate==0.8.1
httplib2==0.19.1
idna==2.8
ifaddr==0.1.6
inflection==0.5.1
ipykernel==5.3.0
ipython==7.15.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jedi==0.17.0
Jinja2==2.11.2
joblib==0.15.1
josepy==1.2.0
json5==0.9.5
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
jupyterlab==2.1.4
jupyterlab-git==0.20.0
jupyterlab-server==1.1.5
keyring==18.0.1
kiwisolver==1.2.0
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
lesspass==10.0.0
lockfile==0.12.2
louis==3.12.0
lxml==4.6.3
macaroonbakery==1.3.1
Mako==1.1.0
manimlib==0.1.11
MarkupSafe==1.1.0
matplotlib==3.2.1
mccabe==0.6.1
mistune==0.8.4
mock==3.0.5
monotonic==1.5
more-itertools==8.6.0
mpmath==1.1.0
multitasking==0.0.9
mypy==0.780
mypy-extensions==0.4.3
Nautilus-scripts-manager==2.0
nbconvert==5.6.1
nbdime==2.0.0
nbformat==5.0.6
nest-asyncio==1.5.1
netifaces==0.10.4
networkx==2.5
nltk==3.5
nose==1.3.7
nose-cover3==0.1.0
notebook==6.0.3
ntlm-auth==1.5.0
numpy==1.18.5
oauthlib==3.1.0
olefile==0.46
opencv-python==4.4.0.46
packaging==20.4
pandas==1.0.5
pandocfilters==1.4.2
paramiko==2.6.0
parsedatetime==2.4
parso==0.7.0
pbr==5.4.5
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
pipx==0.15.6.0
pluggy==0.13.1
ply==3.11
progressbar==2.5
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.17.0
psutil==5.5.1
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.6.2
pycairo==1.16.2
PyChromecast==4.1.0
pycodestyle==2.6.0
pycups==1.9.73
pydocstyle==5.0.2
pydub==0.24.1
pyflakes==2.2.0
Pygments==2.6.1
PyGObject==3.36.0
PyICU==2.4.2
pyinotify==0.9.6
PyJWT==1.7.1
pylint==2.5.3
pymacaroons==0.13.0
PyNaCl==1.3.0
PyOpenGL==3.1.0
pyOpenSSL==19.0.0
pyparsing==2.4.7
PyQt5==5.14.1
pyqtgraph==0.11.0rc0
pyRFC3339==1.1
pyrsistent==0.16.0
PySocks==1.6.8
python-apt==2.0.0+ubuntu0.20.4.5
python-constraint==1.4.0
python-dateutil==2.8.1
python-debian===0.1.36ubuntu1
python-jsonrpc-server==0.3.4
python-language-server==0.33.1
pytz==2019.3
pyxattr==0.6.1
pyxdg==0.26
PyYAML==5.3.1
pyzmq==19.0.1
qiskit==0.23.4
qiskit-aer==0.7.3
qiskit-aqua==0.8.1
qiskit-ibmq-provider==0.11.1
qiskit-ignis==0.5.1
qiskit-terra==0.16.3
qiskit-textbook==0.1.0
qtconsole==4.7.4
QtPy==1.9.0
Quandl==3.6.0
regex==2021.4.4
reportlab==3.5.34
requests==2.22.0
requests-ntlm==1.1.0
requests-oauthlib==1.3.0
requests-toolbelt==0.8.0
requests-unixsocket==0.2.0
retworkx==0.7.2
rope==0.17.0
rsa==4.7.2
scikit-learn==0.23.1
scipy==1.5.0
seaborn==0.11.1
SecretStorage==2.3.1
Send2Trash==1.5.0
simplejson==3.16.0
sip==4.19.21
six==1.14.0
sklearn==0.0
smmap==3.0.4
snowballstemmer==2.0.0
soupsieve==1.9.5
stevedore==3.3.0
sympy==1.6
systemd-python==234
terminado==0.8.3
testpath==0.4.4
threadpoolctl==2.1.0
toml==0.10.1
torbrowser-launcher==0.3.2
tornado==6.0.4
tqdm==4.48.2
traitlets==4.3.3
typed-ast==1.4.1
typing-extensions==3.7.4.2
tzlocal==2.1
ubuntu-advantage-tools==27.0
ubuntu-drivers-common==0.0.0
ufw==0.36
ujson==1.35
unattended-upgrades==0.1
uritemplate==3.0.1
urllib3==1.25.8
usb-creator==0.3.7
userpath==1.4.1
virtualenv==20.0.25
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
wadllib==1.3.3
wcwidth==0.2.3
webencodings==0.5.1
websockets==8.1
Werkzeug==0.16.1
widgetsnbextension==3.5.1
wrapt==1.12.1
xkit==0.0.0
yapf==0.30.0
yfinance==0.1.55
z3-solver==4.8.8.0
zeroconf==0.24.4
zope.component==4.3.0
zope.event==4.4
zope.hookable==5.0.0
zope.interface==4.7.1
